--- a/go/test/endtoend/preparestmt/sVschema.json
+++ b/go/test/endtoend/preparestmt/sVschema.json
@@ -1,0 +1,22 @@
+{
+  "sharded": true,
+  "vindexes": {
+    "xxhash": {
+      "type": "xxhash"
+    }
+  },
+  "tables": {
+    "t1": {
+      "column_vindexes": [
+        {
+          "column": "id",
+          "name": "xxhash"
+        }
+      ],
+      "auto_increment":{
+        "column" : "id",
+        "sequence" : "uks.t1_seq"
+      }
+    }
+  }
+}