--- a/go/test/endtoend/preparestmt/stmt_methods_test.go
+++ b/go/test/endtoend/preparestmt/stmt_methods_test.go
@@ -48,7 +48,7 @@ func TestSelectDatabase(t *testing.T) {
 	require.True(t, rows.Next(), "no rows found")
 	err = rows.Scan(&resultBytes)
 	require.NoError(t, err)
-	assert.Equal(t, string(resultBytes), "test_keyspace")
+	assert.Equal(t, string(resultBytes), "uks")
 }
 
 // TestInsertUpdateDelete validates all insert, update and
@@ -57,7 +57,7 @@ func TestInsertUpdateDelete(t *testing.T) {
 	dbo := Connect(t)
 	defer dbo.Close()
 	// prepare insert statement
-	insertStmt := `insert into ` + tableName + ` values( ?,  ?,  ?,  ?,  ?,  ?,  ?, ?,
+	insertStmt := `insert into vt_prepare_stmt_test values( ?,  ?,  ?,  ?,  ?,  ?,  ?, ?,
 		?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?, ?, ?, ?,
 		?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?,  ?, ?, ?, ?);`
 
@@ -129,7 +129,7 @@ func testReplica(t *testing.T) {
 
 // testcount validates inserted rows count with expected count.
 func testcount(t *testing.T, dbo *sql.DB, except int) {
-	r, err := dbo.Query("SELECT count(1) FROM " + tableName)
+	r, err := dbo.Query("SELECT count(1) FROM vt_prepare_stmt_test")
 	require.Nil(t, err)
 
 	r.Next()
@@ -145,7 +145,7 @@ func TestAutoIncColumns(t *testing.T) {
 	dbo := Connect(t)
 	defer dbo.Close()
 	// insert a row without id
-	insertStmt := "INSERT INTO " + tableName + ` (
+	insertStmt := "INSERT INTO vt_prepare_stmt_test" + ` (
 		msg,keyspace_id,tinyint_unsigned,bool_signed,smallint_unsigned,
 		mediumint_unsigned,int_unsigned,float_unsigned,double_unsigned,
 		decimal_unsigned,t_date,t_datetime,t_datetime_micros,t_time,t_timestamp,c8,c16,c24,
@@ -180,7 +180,7 @@ func TestAutoIncColumns(t *testing.T) {
 // deleteRecord test deletion operation corresponds to the testingID.
 func deleteRecord(t *testing.T, dbo *sql.DB) {
 	// delete the record with id 1
-	exec(t, dbo, "DELETE FROM "+tableName+" WHERE id = ?;", testingID)
+	exec(t, dbo, "DELETE FROM vt_prepare_stmt_test WHERE id = ?;", testingID)
 
 	data := selectWhere(t, dbo, "id = ?", testingID)
 	assert.Equal(t, 0, len(data))
@@ -192,7 +192,7 @@ func updateRecord(t *testing.T, dbo *sql.DB) {
 	// update the record with id 1
 	updateData := "new data value"
 	updateTextCol := "new text col value"
-	updateQuery := "update " + tableName + " set data = ? , text_col = ? where id = ?;"
+	updateQuery := "update vt_prepare_stmt_test set data = ? , text_col = ? where id = ?;"
 
 	exec(t, dbo, updateQuery, updateData, updateTextCol, testingID)
 
@@ -226,7 +226,7 @@ func TestColumnParameter(t *testing.T) {
 	id := 1000
 	parameter1 := "param1"
 	message := "TestColumnParameter"
-	insertStmt := "INSERT INTO " + tableName + " (id, msg, keyspace_id) VALUES (?, ?, ?);"
+	insertStmt := "INSERT INTO vt_prepare_stmt_test (id, msg, keyspace_id) VALUES (?, ?, ?);"
 	values := []any{
 		id,
 		message,
@@ -237,7 +237,7 @@ func TestColumnParameter(t *testing.T) {
 	var param, msg string
 	var recID int
 
-	selectStmt := "SELECT COALESCE(?, id), msg FROM " + tableName + " WHERE msg = ? LIMIT ?"
+	selectStmt := "SELECT COALESCE(?, id), msg FROM vt_prepare_stmt_test WHERE msg = ? LIMIT ?"
 
 	results1, err := dbo.Query(selectStmt, parameter1, message, 1)
 	require.Nil(t, err)
@@ -313,10 +313,7 @@ func TestSelectDBA(t *testing.T) {
 	dbo := Connect(t)
 	defer dbo.Close()
 
-	_, err := dbo.Exec("use uks")
-	require.NoError(t, err)
-
-	_, err = dbo.Exec("CREATE TABLE `a` (`one` int NOT NULL,`two` int NOT NULL,PRIMARY KEY(`one`, `two`)) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4")
+	_, err := dbo.Exec("CREATE TABLE `a` (`one` int NOT NULL,`two` int NOT NULL,PRIMARY KEY(`one`, `two`)) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4")
 	require.NoError(t, err)
 
 	prepare, err := dbo.Prepare(`SELECT
@@ -332,10 +329,10 @@ func TestSelectDBA(t *testing.T) {
 										extra extra,
 										table_name table_name
 									   FROM information_schema.columns
-									   WHERE table_schema = ?
+									   WHERE table_schema = ? and table_name = ?
 									   ORDER BY ordinal_position`)
 	require.NoError(t, err)
-	rows, err := prepare.Query("uks")
+	rows, err := prepare.Query("uks", "a")
 	require.NoError(t, err)
 	defer rows.Close()
 	var rec columns
@@ -441,6 +438,6 @@ func TestBinaryColumn(t *testing.T) {
                   AND column_info.table_schema = ?
                   -- Exclude views.
                   AND table_info.table_type = 'BASE TABLE'
-              ORDER BY BINARY table_info.table_name`, keyspaceName, keyspaceName)
+              ORDER BY BINARY table_info.table_name`, uks, uks)
 	require.NoError(t, err)
 }