--- a/go/test/endtoend/preparestmt/uVschema.json
+++ b/go/test/endtoend/preparestmt/uVschema.json
@@ -1,0 +1,8 @@
+{
+  "sharded": false,
+  "tables": {
+    "sequence_test_seq": {
+      "type":   "sequence"
+    }
+  }
+}