… go/test/endtoend/preparestmt/uSchema.sql