<!--
  Thank you for your contribution to the Vitess project.
  How to contribute: https://vitess.io/docs/contributing/
  Please first make sure there is an open Issue to discuss the feature/fix suggested in this PR.
  If this is a new feature, please mark the Issue as "RFC".
 -->

<!-- if this PR is Work in Progress please create it as a Draft Pull Request -->

## Description

<!-- A few sentences describing the overall goals of the pull request's commits. -->
<!-- If this is a bug fix and you think the fix should be backported, please write so. -->
Added a benchmark test for prepared statements covering insert, select, update, and delete operations. Utilized this test to analyze heap allocations, revealing that approximately 20% of allocations are consumed in generating a query plan, which can be avoided during prepared statement execution.

![image](https://github.com/user-attachments/assets/d190ebd6-5b6d-462b-8264-9d83dd9e6073)

```
goos: darwin
goarch: arm64
pkg: vitess.io/vitess/go/test/endtoend/preparestmt
cpu: Apple M1 Max
                      │ /Users/planetscale/benchmark/prepare-no-change.txt │
                      │                       sec/op                       │
PreparedStmt/Insert-4                                          490.9µ ± 2%
PreparedStmt/Select-4                                          305.8µ ± 0%
PreparedStmt/Update-4                                          342.2µ ± 2%
PreparedStmt/Delete-4                                          311.3µ ± 1%
geomean                                                        355.6µ
```

Implemented an optimization to bypass query parsing and normalization for already-prepared statements. Benchmark results demonstrated significant query execution performance gains:
* 6.40% improvement in query execution time for basic select queries
* 9.08% improvement in query execution time for join-based select queries
* 3.62% improvement in query execution time for delete queries
* No statistically significant change in query execution time for insert and update queries

The overall geometric mean improvement in query execution across all operations was **4.94%**, with potential for further gains as query size increases.

```
goos: darwin
goarch: arm64
pkg: vitess.io/vitess/go/test/endtoend/preparestmt
cpu: Apple M1 Max
                                        │ nochange-2s-v3.txt │         prepared-2s-v3.txt         │
                                        │       sec/op       │    sec/op     vs base              │
PreparedStmt/Insert-8                            506.5µ ± 9%   490.1µ ± 14%       ~ (p=0.065 n=6)
PreparedStmt/Select-8                            306.7µ ± 5%   287.1µ ±  5%  -6.40% (p=0.002 n=6)
PreparedStmt/Join_Select:Simple_Route-8          300.5µ ± 4%   273.3µ ±  3%  -9.08% (p=0.002 n=6)
PreparedStmt/Update-8                            351.0µ ± 5%   343.2µ ±  3%       ~ (p=0.240 n=6)
PreparedStmt/Delete-8                            318.9µ ± 4%   307.3µ ±  4%  -3.62% (p=0.015 n=6)
geomean                                          349.6µ        332.4µ        -4.94%
```

## Related Issue(s)

<!-- List related issues and pull requests. If this PR fixes an issue, please add it using Fixes #????  -->

## Checklist

-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on CI?
-   [x] Documentation was added or is not required