@harshit-gangal 2025-01-03
A follow-up PR will include the implementation and final performance improvement results.

Temporary changes from [this commit](https://github.com/planetscale/vitess/commit/275743ba2774d096abf3eba3c00c52bd30a62f12) were used to evaluate the potential improvement.